feat: allow bots to plunder resources (#433)
* const

* deps

* deps

* err

* derive_const

* get_opinion

* deps

* format

* lint fix

* diplomacy

* derive_const

* plunder

* power ops

* minimumReleaseAge

* .taurignore

* deps

* derive_const

* const

* const

* const

* const

* personnel_of

* deps

* const

* derive

* ConstDeref

* f64

* move

* power

* move files

* feat: allow bots to plunder resources from other rulers

* u32

* get_world_personnel

* deps

* fmt

+++ crates/nil-server-types/src/world.rs
@@ -21,7 +21,7 @@ pub struct RemoteWorld {
   pub has_password: bool,
   pub current_round: RoundId,
   pub round_duration: Option<RoundDuration>,
-  pub active_players: usize,
-  pub total_players: usize,
+  pub active_players: u32,
+  pub total_players: u32,
   pub continent_size: ContinentSize,
 }
